+++ CHANGELOG.md
@@ -17,10 +17,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Support for User Profile properties upload [gautamdsheth]
 
 ### Changed
+
 - Deprecated ApplyProvisioningHierarchy and replaced with ApplyTenantTemplate
 - IsProvisioningComplete waits for max 20 minutes and logs a warning instead of throwing an exception when a longer wait is needed
-- Fix for reading/setting search redirect URL on a root site
-- Fix for handling SharePoint groups and groups' owners [NicolajHedeager]
+- Fix: reading/setting search redirect URL on a root site
+- Fix: Updated handling SharePoint groups and groups' owners #2444 [NicolajHedeager]
+- Fix: Update us gov auth endpoint #2463 [gobigfoot]
 
 ## [3.14.1910.1 - October 2019 intermediate release]